Refactor Mesh feature.
