+++ demo/src/features/mesh/mod.rs
@@ -1,172 +1,94 @@
-use distill::loader::handle::Handle;
-use rafx::base::slab::{DropSlab, DropSlabKey};
-use rafx::framework::{MaterialPassResource, ResourceArc};
-use rafx::nodes::RenderView;
-use rafx::nodes::{FrameNodeIndex, GenericRenderNodeHandle, RenderNodeCount, RenderNodeSet};
-
-mod extract;
-
-mod prepare;
-
-mod write;
-
-mod plugin;
-pub use plugin::MeshRendererPlugin;
-
-pub mod shadow_map_resource;
-
-use rafx::framework::DescriptorSetArc;
-use write::MeshCommandWriter;
-
-const PER_VIEW_DESCRIPTOR_SET_INDEX: u32 =
-    shaders::mesh_frag::PER_VIEW_DATA_DESCRIPTOR_SET_INDEX as u32;
-const PER_MATERIAL_DESCRIPTOR_SET_INDEX: u32 =
-    shaders::mesh_frag::PER_MATERIAL_DATA_DESCRIPTOR_SET_INDEX as u32;
-const PER_INSTANCE_DESCRIPTOR_SET_INDEX: u32 =
-    shaders::mesh_frag::PER_OBJECT_DATA_DESCRIPTOR_SET_INDEX as u32;
-
-use crate::assets::gltf::MeshAsset;
-use crate::components::{
-    DirectionalLightComponent, PointLightComponent, PositionComponent, SpotLightComponent,
-};
-pub use shaders::depth_vert::PerObjectDataUniform as ShadowPerObjectShaderParam;
-pub use shaders::depth_vert::PerViewDataUniform as ShadowPerViewShaderParam;
-pub use shaders::mesh_frag::PerObjectDataUniform as MeshPerObjectFragmentShaderParam;
-pub use shaders::mesh_frag::PerViewDataUniform as MeshPerViewFragmentShaderParam;
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-pub enum LightId {
-    PointLight(legion::Entity), // u32 is a face index
-    SpotLight(legion::Entity),
-    DirectionalLight(legion::Entity),
-}
-
-#[derive(Clone)]
-pub enum ShadowMapRenderView {
-    Single(RenderView), // width, height of texture
-    Cube([RenderView; 6]),
-}
-
-pub struct ExtractedDirectionalLight {
-    light: DirectionalLightComponent,
-    entity: legion::Entity,
-}
-
-pub struct ExtractedPointLight {
-    light: PointLightComponent,
-    position: PositionComponent,
-    entity: legion::Entity,
-}
-
-pub struct ExtractedSpotLight {
-    light: SpotLightComponent,
-    position: PositionComponent,
-    entity: legion::Entity,
-}
-
-pub fn create_mesh_extract_job() -> Box<dyn ExtractJob> {
-    Box::new(ExtractJobImpl {})
-}
-
-//
-// This is boiler-platish
-//
-pub struct MeshRenderNode {
-    pub mesh: Option<Handle<MeshAsset>>,
-    pub transform: glam::Mat4,
-}
-
-#[derive(Clone)]
-pub struct MeshRenderNodeHandle(pub DropSlabKey<MeshRenderNode>);
-
-impl MeshRenderNodeHandle {
-    pub fn as_raw_generic_handle(&self) -> GenericRenderNodeHandle {
-        GenericRenderNodeHandle::new(
-            <MeshRenderFeature as RenderFeature>::feature_index(),
-            self.0.index(),
-        )
-    }
-}
-
-impl Into<GenericRenderNodeHandle> for MeshRenderNodeHandle {
-    fn into(self) -> GenericRenderNodeHandle {
-        self.as_raw_generic_handle()
-    }
-}
-
-#[derive(Default)]
-pub struct MeshRenderNodeSet {
-    meshes: DropSlab<MeshRenderNode>,
-}
-
-impl MeshRenderNodeSet {
-    pub fn register_mesh(
-        &mut self,
-        node: MeshRenderNode,
-    ) -> MeshRenderNodeHandle {
-        MeshRenderNodeHandle(self.meshes.allocate(node))
-    }
-
-    pub fn get_mut(
-        &mut self,
-        handle: &MeshRenderNodeHandle,
-    ) -> Option<&mut MeshRenderNode> {
-        self.meshes.get_mut(&handle.0)
-    }
-
-    pub fn update(&mut self) {
-        self.meshes.process_drops();
-    }
-}
-
-impl RenderNodeSet for MeshRenderNodeSet {
-    fn feature_index(&self) -> RenderFeatureIndex {
-        MeshRenderFeature::feature_index()
-    }
-
-    fn max_render_node_count(&self) -> RenderNodeCount {
-        self.meshes.storage_size() as RenderNodeCount
-    }
-}
-
 rafx::declare_render_feature_mod!();
 rafx::declare_render_feature_renderer_plugin!();
+
 rafx::declare_render_feature!(MeshRenderFeature, MESH_FEATURE_INDEX);
 
-pub struct ExtractedFrameNodeMeshData {
-    world_transform: glam::Mat4,
-    mesh_asset: MeshAsset,
+mod extract;
+mod prepare;
+mod write;
+
+mod public;
+
+use distill::loader::handle::Handle;
+use legion::Resources;
+use rafx::assets::MaterialAsset;
+use rafx::nodes::{FramePacketBuilder, RenderView, RenderViewSet};
+use rafx::visibility::{DynamicVisibilityNodeSet, StaticVisibilityNodeSet};
+
+pub use public::*;
+
+pub struct RendererPluginImpl;
+
+struct StaticResources {
+    pub depth_material: Handle<MaterialAsset>,
 }
 
-impl std::fmt::Debug for ExtractedFrameNodeMeshData {
-    fn fmt(
+impl RendererPlugin for RendererPluginImpl {
+    fn configure_render_registry(
         &self,
-        f: &mut std::fmt::Formatter<'_>,
-    ) -> std::fmt::Result {
-        f.debug_struct("ExtractedFrameNodeMeshData")
-            .field("world_transform", &self.world_transform)
-            .finish()
+        render_registry: RenderRegistryBuilder,
+    ) -> RenderRegistryBuilder {
+        render_registry.register_feature::<MeshRenderFeature>()
+    }
+
+    fn initialize_static_resources(
+        &self,
+        asset_manager: &mut AssetManager,
+        asset_resource: &mut AssetResource,
+        _extract_resources: &ExtractResources,
+        render_resources: &mut ResourceMap,
+        _upload: &mut RafxTransferUpload,
+    ) -> RafxResult<()> {
+        let depth_material =
+            asset_resource.load_asset_path::<MaterialAsset, _>("materials/depth.material");
+
+        asset_manager.wait_for_asset_to_load(&depth_material, asset_resource, "depth")?;
+
+        render_resources.insert(StaticResources { depth_material });
+
+        render_resources.insert(ShadowMapResource::default());
+
+        Ok(())
+    }
+
+    fn add_render_views(
+        &self,
+        extract_resources: &ExtractResources,
+        render_resources: &RenderResources,
+        render_view_set: &RenderViewSet,
+        frame_packet_builder: &FramePacketBuilder,
+        static_visibility_node_set: &mut StaticVisibilityNodeSet,
+        dynamic_visibility_node_set: &mut DynamicVisibilityNodeSet,
+        render_views: &mut Vec<RenderView>,
+    ) {
+        let mut shadow_map_resource = render_resources.fetch_mut::<ShadowMapResource>();
+        shadow_map_resource.recalculate_shadow_map_views(
+            &render_view_set,
+            extract_resources,
+            &frame_packet_builder,
+            static_visibility_node_set,
+            dynamic_visibility_node_set,
+        );
+
+        shadow_map_resource.append_render_views(render_views);
+    }
+
+    fn add_extract_jobs(
+        &self,
+        _extract_resources: &ExtractResources,
+        _render_resources: &RenderResources,
+        extract_jobs: &mut Vec<Box<dyn ExtractJob>>,
+    ) {
+        extract_jobs.push(Box::new(ExtractJobImpl::new()));
     }
 }
 
-pub struct PreparedSubmitNodeMeshData {
-    material_pass_resource: ResourceArc<MaterialPassResource>,
-    per_view_descriptor_set: DescriptorSetArc,
-    per_material_descriptor_set: Option<DescriptorSetArc>,
-    per_instance_descriptor_set: DescriptorSetArc,
-    // we can get the mesh via the frame node index
-    frame_node_index: FrameNodeIndex,
-    mesh_part_index: usize,
+// Legion-specific
+
+pub fn legion_init(resources: &mut Resources) {
+    resources.insert(MeshRenderNodeSet::default());
 }
 
-impl std::fmt::Debug for PreparedSubmitNodeMeshData {
-    fn fmt(
-        &self,
-        f: &mut std::fmt::Formatter<'_>,
-    ) -> std::fmt::Result {
-        f.debug_struct("PreparedSubmitNodeMeshData")
-            .field("frame_node_index", &self.frame_node_index)
-            .field("mesh_part_index", &self.mesh_part_index)
-            .finish()
-    }
+pub fn legion_destroy(resources: &mut Resources) {
+    resources.remove::<MeshRenderNodeSet>();
 }
