+++ demo/src/features/mesh/public/mod.rs
@@ -1,0 +1,5 @@
+mod mesh_render_node_set;
+pub use mesh_render_node_set::*;
+
+mod shadow_map_resource;
+pub use shadow_map_resource::*;
